rename helperFunctions to helpers
